key-transformer: Instead of trying to be clever, just use a big list
Since we're not in the firmware where we need to save as much space as possible,
we can just use a big list with 64k elements for each keycode. Simplified both
lookup and serialization.

Since not all codes are useful, and we want grouping, this patch sets things up
accordingly.

Partly addresses #18.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>