Simplify writing to the Serial port
For Windows and Linux, we do not need to do the request splitting dance. The
splitting doesn't even work on Windows properly, because despite best efforts,
they end up overlapping.

So we're back to just sending the request as-is, in one big write. If this
breaks OSX, we'll figure something out - perhaps platform-specific code will be
the answer.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>