Completely revamped flashing flow
When clicking the "Upload" button on the flashing card, instead of flashing
right away, a dialog will pop up with some explanatory text about what will
happen, and what to expect. This fixes #83.

After flashing is done, we'll fall back to the keyboard selection screen,
because reconnecting is not reliable yet. This should fix the white screen after
flashing, and as such, this also closes #69.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>