main: Catch the gpu-process-crashed event, and print a message
If, for some reason, the renderer process crashes, we can't display a proper
error message in the browser window, so log one to the terminal with a few
instructions about how to file a useful bug report about it.

This does not fix #112, but in the event of a similar bug occurring, we'll have
more clue and more information to start with.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>