KeySelector: Major redesign
Instead of having a good few number of different ways to lay out the selectable
keys on the right side of the keymap, move it to the bottom, and present the
list as a simple list of buttons one after another. The only exception are
unknown keycodes, those are still presented as a numeric input field.

This fixes #120.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>