Support FirmwareUpdate & Welcome for non-Kaleidoscope firmware
We support a number of keyboards that do not come with Kaleidoscope by default,
and their factory firmware doesn't have a virtual serial port enabled either. We
weren't able to find these previously, making it that much harder to try
Chrysalis with a supported device.

With this change, we do a second pass of search, for keyboards that aren't
Kaleidoscope-powered out of the factory, and find them via the `usb` library.
For these boards, we do not really "connect" to them, just set things up well
enough for the `FirmwareUpdate` and `Welcome` screens to work. This in turn
allows one to flash a Kaleidoscope-based firmware onto these boards, and try
Chrysalis.

We do make an assumption that all connected keyboards are unique, that we do not
have two of the same keyboard attached at the same time.

Fixes #106.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>