KeySelector: Add support for moddable keys
When encountering a modded key (`LSHIFT()` & friends), display the un-modded
group. In the key list, look at the keycode, and handle modifiers. When
selecting a key via the buttons, apply the selected modifiers too.

In practice, this means we can toggle the modifiers on and off now, and this
fixes #63.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>