LayoutEditor: Use a contextual AppBar for selecting the default layer
Instead of a dropdown on the right side of the toolbar, offer an Edit button in
the same position, which switches the AppBar into a contextual mode. In this
mode, the page menu becomes a Close/Cancel button, the Edit becomes Save, and
selecting a layer will select it as the new default (not saved until Save is
pressed).

Fixes #108.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>