Welcome: Implement a "Reconnect" button
If we're connected to a keyboard with a serial port, but did not find Chrysalis,
there's a small chance that we connected while the keyboard was booting up, or
while something else was reading from the port too. In this case, offer an easy
way to reconnect.

If we're connected to a board without a serial port, we do not offer the
reconnect.

Fixes #157.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>