Implement layer copying
The ColormapEditor and LayoutEditor screens now have a "more menu", where one
can clear the layer, or copy it to another. Clearing is only available for
non-read-only layers, and only writeable layers are offered as copy targets.
Once copied, we switch there too. The changes are not immediately saved,
clicking the "Save Changes" button is still required.

Fixes #110.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>