chrysalis-flash: Split up Avr109, and allow configurable timeouts
Split up Avr109 into two parts: one that does the reset, and another that does
the flashing. This is so that if we have a device already in bootloader mode,
we'd still be able to flash it, without having to reset it again.

While there, introduce configurable timeouts with nice names, and bump the time
we allow the bootloader to come up to 2s. The names fix #15.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>