Add support for the Keyboardio Model01 in boot mode
By upgrading the hardware libraries, we have access to `Model01Bootloader`, a
variant of the `Model01` descriptor. This allows Chrysalis to detect such a
Model01, and allows us to flash it, without needing to boot a user firmware for
Chrysalis to recognise it.

Fixes #240, and since the hardware library also bumped the time it waits for the
bootloader to come up, this also fixes #238.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>