travis: Disable filter_secrets
Disable `filter_secrets`, which hopefully makes our windows builds work. In
general, this wouldn't be a safe thing, because a malicious pull request could
leak our secrets. But we only run builds for master, not for pull requests or
any other branch, so what we do is well controlled, and shouldn't leak any.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>