Move a few less-than-common settings from LayoutEditor to Settings
Now that we copy the built-in layers to EEPROM when we encounter an
uninitialized EEPROM, having the default layer selector on the layer bar is not
needed. Moved it to the Settings page. Fixes #288.

Similarly, selecting whether to use custom layers only or not belongs in
Settings too. Fixes #286.

Likewise, whether to show or hide the default / built-in layers is also a
setting, and belongs there. Since this is not saved on the keyboard, but we want
to persist the setting, we're using `electron-settings` to store the setting on
disk. Fixes #287 and fixes #277.

The moved settings do not take effect immediately like they used to, one needs
to use the new "Save changes" button on the page. Fixes #285.

This also fixes #194, due to having figured out how to store settings.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>