Editor: Fix editing the colormap on layer 0
When deciding whether to set the colors, we wanted to avoid setting it for
negative layers (ie, when using custom layers only, but having default layers
shown, and being on a default layer). However, the comparison was done wrong, we
compared `layer > 0` instead of `layer >= 0`. This fixes that, and setting
colors on the base layer works again now.

Fixes #352.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>