ColorPalette: If disabled, don't show the palette
If the palette is disabled, don't show it at all. This prevents an error when
the palette is used for negative layers. The proper fix here would be to show
the palette in that case too, but in a disabled state. But that's more work, so
to prevent the error, do the easy fix now, and the proper one later.

Signed-off-by: Gergely Nagy <me@gergo.csillger.hu>