Upgrade Electron and SerialPort
Upgrade to the latest `@chrysalis-api` packages, which pull in the new
SerialPort. We also upgrade from Electron 4 to Electron 7.

For both of these to work, we transition from `port.comName` to `port.path`, and
explicitly enable `nodeIntegration` in the renderer process (which was the
default in Electron 4, but no longer is in Electron 7).

Signed-off-by: Gergely Nagy <algernon@keyboard.io>