tools/get-artifact: Be more verbose, and URL-encode the version
We need to encode the URL ourselves, as curl expects an urlencoded URL, and
won't do it for us. We only need to treat `+` specially, so I went with a very
simple way to do that.

While there, also made the script a little bit more verbose, so we know what it
is doing, and where it is saving the downloaded file to.

Signed-off-by: Gergely Nagy <me@gergo.csillger.hu>