Disable the sandbox on the AppImage build
Disable the sandbox automatically in the AppImage build. This is done by running
the prod build through a wrapper that detects whether we can enable it, and
disables it if we can't.

This is only done on Linux, and only in the prod build: `yarn run start` will
still require an explicit `--no-sandbox`.

Signed-off-by: Gergely Nagy <me@gergo.csillger.hu>