Set `NODE_ENV` correctly in the main process
Due to electron-userland/electron-webpack#275, we need to set `NODE_ENV` in the
main process, as early as possible, to the correct value. We also need to do
that in a strange, perhaps a bit ugly and verbose way, to avoid webpack
optimizing important parts of it out.

With this workaround, we can use the latest webpack, without serial operations
getting stalled.

Signed-off-by: Gergely Nagy <me@gergo.csillger.hu>