api/flash/Avr109: Major refactor
Refactor the flashing process so that we're using async/await more, instead of
nested callbacks. This makes the function shorter, flatter, easier to follow.

While there, if we do not find a bootloader, throw an error instead of
proceeding with a flashing attempt. Furthermore, we now call the `callback`
function with the correct step names.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>