Add a LoadingScreen component
When communicating with the keyboard, and we expect that it might take a while,
try showing a loading animation instead of just a partial render. This should
result in a considerably nicer user experience.

The component is currently only used by the Editor screen, but we may want to
use it elsewhere too.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>