Editor/Collapsible: Stop changing the expanded state on props change
Doing so resulted in some annoying behaviour when the expanded state of boxes
were changed after changing the keycode to one in a different section, while
having multiple boxes open at the same time.

We just set the initial state on first render, and then leave it up to the user
to expand/collapse as they see fit, without second guessing.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>