github/workflows: Set the build script name explicitly
By default, the action we use defaults to `build` as the build script name,
while we use `compile` for the same purpose. I believe `compile` is the better
name, so tell the action to use that.

We need this step so that our sources are compiled before they're packed up.
Without compilation, many of the things we want to end up in the package simply
won't be there, and the build would fail.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>