api/keymap: Teach withModifiers() about layered labels
A while ago, we made it possible to have more fine grained control over how
labels are displayed at various key sizes. This turned some key labels from
simple strings to objects. However, `withModifiers()` wasn't taking this into
account, and always treated the label as a string. This resulted in situations
where we had `A+[object object]` as labels.

This patch updates `withModifiers` to work with the layered system, and display
all labels correctly. We also moved the modifiers to a hint, to make the display
slightly nicer.

Fixes #645.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>