Editor: Fix populating an empty custom keymap with the default
When the Editor screen renders, we scan the custom keymap, to see if it is
empty, and if it is, we copy the default keymap over, so that on fresh
keyboards, Chrysalis will display the built-in layout instead of an empty
keymap.

However, with the database redesign in 0.8, a property changed names, and the
code that does the emptyness check was using the old property name, and the
check thus always failed.

We simply change the property name to the correct one, and populating the custom
keymap will work once again.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>