api/keymap: Fix the dualuse database entries
The way we add dualuse keys to the database is that we iterate over the base
keyset, and augment those. However, since these loops were written, we updated
the base keyset, and they contain modifier-augmented variants too. We do _not_
want to add dual-use augmentation on top of that. That totally messes up the
database.

As such, restrict the keys we augment to those that have a keycode between 0 and
255.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>