api/keymap: Correctly load shifted and altgr'd labels
When loading a keyboard layout from CLDR, we were setting the `label.shifted`
property of the base key, but we did not touch the entry for the shifted/altgr'd
keycode. This resulted in a display issue where if we wanted to display the base
key _and_ its modified symbol, we'd do that properly, because that's looked up
via the base keycode. However, when we wanted to display the shifted symbol,
that's looked up via the shifted keycode, so we need to update that entry in the
database  too.

This fixes the display part of #513.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>