CI: Bump the node versions used for CI from 10/12 to 16
For historical reasons, we've been using node 10 for both release and snapshot
builds, and 12 for linting. As that's rather old, and some of our transitive
dependencies no longer support anything older than 12, lets move forward, and
bump node to 16. That's the current LTS release, and our whole stack should work
well with it, Electron and serialport included.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>