stylelint: Lint css files only, and fix keymap.css
Stylelint doesn't have good support for JSS, so we can't properly check our JS
files. As such, restrict the linting to css files only. For that to work, we
need to unquote the arguments passed to stylelint, otherwise the shell will not
expand the wildcards.

With the call fixed, stylelint found an issue in `src/styles/keymap.css`, which
this patch fixes too.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>