components/Header: Make the board button visible again
With the recent change to theming, the board button ended up with the same color
as the background, making it effectively invisible. With this change, we
explicitly set its color to inherit from its parent, the toolbar, so it will
have the same color as the rest of the header.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>