api/keymap: When switching layouts, update the base labels for DualUse keys
When we switch layouts, we were updating the base labels only, but not any
dualuse keys, which resulted in labels that were incorrect for non-US QWERTY
layouts for keys with secondary actions.

With this change, we now update both modifier and layer secondary functions with
the proper labels from the new layout.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>