screens/FirmwareUpdate: Make it possible for shipped firmware to have non-hex type
When looking for firmware files shipped with Chrysalis, we were looking for
files with a `.hex` extension. The Model100 uses `.bin`, so we should be looking
for that instead. To make this possible, we introduce a `firmwareType` property
to the hardware descriptor's `info` section, and use that instead of `hex` if
present. If not present, we'll simply fall back to `hex`.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>