CI: When deleting the snapshot release, do so locally too
As part of our snapshot workflow, we delete and re-create the snapshot
release (and tag), to have a consistent tag to link to, and upload artifacts to.
However, when doing so, we only deleted the release and the tag remotely, not on
the local checkout.

This led to every second build failing, until the publish step was re-run.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>