Preferences: Reset the keyboard after doing an eeprom clear
After clearing EEPROM, we need to reset the keyboard so that all the plugins
have a chance to act accordingly. Not doing so leaves the keyboard in a very
confused state, and things *will* break.

To reset a device, we first try the `device.reset` command, and fall back to
doing a 1200 baud serial HUP. We do the latter unconditionally, to make sure
that either one of them gets the job done.

Worst case, we'll reset twice.

Fixes #739.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>