Editor: Do not allow augmenting a modifier with itself
When configuring a modifier key, do not allow augmenting it with itself. We
already disallowed this when the base keycode was the modifier, but we need to
disallow the case where the current code is the modifier in question, too.

Fixes #705 by not allowing the situation to occur in the first place.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>