i18n: Fix the update instruction refreshing
First of all, we need to call `i18n.refreshHardware` in the case of a successful
connection too, not only when the keyboard is a non-Serial one.

Second, due to the recent restructuring that resulted in
`port.focusDeviceDescriptor` becoming a thing, we no longer need to destructure
the arguments passed to `i18n.refreshHardware`, it now receives the object it
needs directly.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>