KeyboardSelect: Do not show the linux permission warning with no keyboard selected
The permission check only makes sense when we have a keyboard selected. If we do
not - likely because there aren't any supported keyboards connected -, then we
should not display the warning.

In the `foo && foo.bar` => `foo?.bar` cleanup, this ended up getting broken, so
I added a note to avoid that in the future.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>