api/flash & api/hardware: Reconnect before restoring EEPROM
Before attempting to restore EEPROM, wait for the keyboard to come back, and
reconnect. Otherwise we have nothing to restore EEPROM to.

Luckily, we already had `focus.waitForSerialBootloader`, which I could repurpose
to wait for any serial device. We then just wrap that, and connect after the
wait's done, and we're golden.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>