api/flash: With structural EEPROM save/restore, we do not need the final reboot
We needed the final reboot because we used to use `eeprom.contents` to restore
EEPROM, and that did not tell the plugins on the firmware that settings changed,
so they weren't aware. We forced a reboot so that plugins pick the changes up.

However, since we use structural save/restore, the plugins are aware now,
without the reboot. As such, remove this step.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>