api/focus: Use `focus.command` to save & restore the eeprom
We want to re-use the saved information for purposes other than immediately
restoring - like saving the backup to a file for recovery in case of errors.

For these use cases, it is more beneficial to have the data processed. As such,
use `focus.command` rather than `focus.request` to save and restore them, so we
get the processing bits too.

Adjusted the list of slots to save and restore accordingly.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>