Editor: Teach LayoutSharing about the eeprom-backup file format
When flashing, the process will now save a backup of the structured eeprom
contents. That's a little different than what LayoutSharing exports, so convert
it quickly on load.

We still load & save the format we used, mind you, because we do not have enough
data to properly match the eeprom-backup variant. Thankfully, the load-time
conversion is just a matter of deciding which key to read data from. The data
itself is 100% the same, just the key differs.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>