KeyboardSelect: Restore the auto-connect functionality
This functionality was lost when I merged master into the branch, because the
way to do it became very different.

The way we do it now is that we use a new state: `tryAutoConnect`. This starts
as true, so on boot, we auto-connect. When disconnecting, we explicitly set it
to false, though, so we do not automatically reconnect.

If we navigate away and come back, we'll still go and auto-connect, as in the
original implementation.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>