api/focus: Do not send unknown requests
Replaced `focus.probe()` with `focus.supported_commands()` which returns the
list of supported commands - the output of `help`. We cache this value for the
remainder of the connection, and clear it on `close` or an `open`.

This lets us avoid sending requests that the keyboard doesn't implement,
reducing the traffic between keyboard and host. We do this by checking the same
list in `focus.request()`, and turning the request into a no-op if the command
is unsupported.

If the list is uninitialized, we'll send the request anyway.

All of the above made it possible to remove similar code from
`@render/ActiveDevice`. We prefer doing this in `focus`, at a lower level, so
everything benefits from it, not just things that build on `ActiveDevice.

Fixes #802.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>