Editor: Dispatch onKeyChange based on the macro editor
If we have the macro editor open, we want to change the currently edited
step. If we don't, we want to change the one on the keymap. To this end,
split `onKeyChange` into two parts, one for the macro editor, one for
the keymap.

For the various selectors (the floating key picker and the sidebar) to
work, we also need to track which key is currently selected, so that
they see "inside" a macro, so to say. The new `selectorKey` state tracks
just that.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>