api/macros: Serialize should fill all the available space
When serializing, fill the remaining space with 255s, to clear the storage of
any leftovers. This prevents us from misparsing the macro area after we reduced
the size of our macros.

However, teach `macros.serialize` to do the filling optionally, so we can
serialize without filling when checking how much space we use.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>