Editor: Adjust the Sidebar to work with the MacroEditor too
We want to use the sidebar to update "tap" macro steps. The Editor
already changed the `onKeyChange` function to update macro steps when
the macro editor is open, but the sidebar is still doing a lot of
keycode lookups itself: we pass it a keymap, a layer, and an index, and
it looks up the keycode from that.

If we want to use the sidebar to assign keys to "Tap" macro steps, we
can't have it do these kinds of lookups. So this patch changes all the
relevant components to use the `currentKey` prop that is being passed
along to the Sidebar, and contains the current, already looked up key,
be it on the keymap, or in a macro step.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>