FirmwareUpdate: Let the user know when they can release PROG
When we reach the waiting-for-bootloader step of the firmware flashing process,
it is safe to release the PROG key (or its equivalent). Pop up a toast letting
the user know.

We do this when we start to wait for the bootloader, rather than after, to give
enough time for the user to react. If we do it at the start of `flash` (the
usual next step), we only have a couple of seconds to react before flashing
finishes. Doing it at this time gives us a few more.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>