FirmwareUpdate: Auto-hide the success toast after 10 seconds
It's great to notify the user about a successful firmware upgrade, but don't
leave the toast hanging indefinitely. A 10s popup is long enough to be noticed,
but short enough to not be annoying.

The idea here is that after a flash, the user may want to connect back to the
keyboard - or the auto-connect might do it for them anyway -, and the toast in
that case is in the way. It's better if they don't have to explicitly close it.

We keep the error toast up indefinitely, however, because that does require
human attention.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>