SideBar: Hide parts for which we do not have the plugin installed for
The goal is to hide - unconditionally for now, optionally later - functionality
from the sidebar for which we do not have the plugin installed for
firmware-side. This reduces clutter, and avoids surprises where a user might
assign a functionality to a key without the plugin being present, and then left
wondering why it doesn't work.

We achieve this by having a new `usePluginVisibility` hook, which checks if the
active device has the appropriate plugin available. When we know the visibility,
we can decide whether to display a piece of functionality or not.

For most sections, we can hide the entire section. In other cases (like
Modifiers and LayerKeys) we only disable parts of the functionality, that
require a particular plugin on the firmware.

For now, `usePluginVisibility` only checks the plugin list. In a follow-up
patch, we'll teach it to check a setting too, which can force visibility on,
even if the plugin is absent.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>