Preferences: Add an option to show all available key types in the Editor
With the previous commit, we started hiding functionality from the Editor when
the required plugin is missing on the firmware side. There are, however,
scenarios where we do want to configure all kinds of keys, even those
unsupported by the firmware.

The new option lets us do so, and `usePluginVisibility` adheres to the setting.

Together with the previous commits, this fixes #836.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>