Upgrade `usb` from 2.3.1 to 2.4.2.
Previously, we were using `usb` 2.3.1, but in a recent spree of regenerating
`yarn.lock`, we ended up resolving the version to be installed to 2.4.1.
Unfortunately, `usb` 2.4.0 made the package context aware (a good thing), but
also broke it on Windows completely (not such a good thing).

Luckily, `usb` 2.4.2 was released some 21 hours ago, which fixes the regression
on Windows, and using it, Chrysalis can connect to keyboards on Windows again.

Fixes #849.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>