tools/preinstall.js: Don't rewrite package.json unnecessarily
The goal of this preinstall script is to add a build number to the package
version within our CI environment. Outside of CI, it should do nothing. However,
it was unconditionally overwriting `package.json`, with a version that has no
trailing newline.

That's an issue, because most editors will add one, so running `yarn install`
always removed that, ending up with the change making the git working directory
dirty.

As there is no reason to save the file when we're not running in the CI
environment, move the writing inside the if branch. We don't care if the working
tree becomes dirty in CI, it will be discarded anyway. Not running
unconditionally will stop newlines getting removed, and make development that
much less annoying.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>