package.json: Adjust the async resolution for some packages
An unused dependeny of avrgirl-arduino depends on an ancient `async` package
Dependabot keeps complaining about. Adjust the dependency to a fixed version.

As we're not using the parts of AvrGirl that make use of this particular version
of async, there should be no issues. The parts we use were already using the 2.6
series.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>