Editor: Buttonize the layer keys
Instead of having a second dropdown for the target layer, render a set of
buttons below the type chooser. This makes choosing a layer key faster, and the
user experience more consistent.

The buttons are always in the same position, and are always visible, regardless
of the type chosen.

Fixes #752.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>