Improve our theme system
Instead of having a toggle between light and dark mode, support arbitrary themes
by storing the theme name in a setting, rather than a dark mode toggle.

This allows us to have a "system" theme (along with "light" and "dark"), which
follows the system theme on operating systems that have one (macOS & Windows).

The old `ui.darkMode` setting is honored, and will be migrated automatically to
the new setting. For new users, we default to the system theme.

Fixes #874.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>