api/focus: Temporarily bump the timeout to 30s
We're seeing disconnects on Windows when using Chrysalis with the Model 100.
From the debug logs, it looks like there is communication between the keyboard
and the host, but after a `keymap.custom` request, we end up back at the
keyboard selection screen.

That suggests we're triggering a timeout, so increasing it here will hopefully
help.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>