Preferences: Use a slightly darker body background
We want to make a few elements stand out a bit more, so we're setting a slightly
darker (or in case of dark mode, slightly lighter) background on body. We set it
on the body so it will be the default background, and we won't have to go to
great lengths to hide the original body background.

Because we need the background color  to be theme-dependent, we make use of the
`react-helmet` library, which helps us set properties on elements normally
outside of React's reach - like `body`.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>