Preferences: Major architectural rework of MyKeyboard
Rather than MyKeyboard having all the knowledge and state, and pushing it down
to the components that render the tab, have a list of commands to run at save
time, and push a single handler down to children, which allows us to register
new command + args pairs.

Upon save, we run those focus commands, with the necessary data, and then clear
the queue. On cancel, we'll clear our queue.

This makes it considerably easier to add support for new plugins, as there are
less files to modify, and knowledge about a particular plugin is closer to the
place that displays it: in the same component, rathern than two components
higher in the tree, at least.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>