Preferences: Document a perhaps surprisingly way we handle a property
In both PreferenceSection and PreferenceSwitch, we handle the `loaded` property
in a perhaps surprising way: we first check if it is undefined, and treat that
as if the property were true, otherwise we use the property as it is.

We do this because we want both of these components to be usable in scenarios
where loading makes no sense, without having to specify a `loaded={true}`
property.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>