Preferences: Lift out a common pattern into a helper component
A fair number of our components on the Preferences screen were using a pattern
that had a box, a heading or two, and a simple input on the right side of the
component.

These were extracted into `PreferenceWithHeading`, and all users updated to
prefer this over just copy & pasting the same pattern.

Thanks @obra for noticing the pattern!

Signed-off-by: Gergely Nagy <algernon@keyboard.io>