api/focus: Add constants to our keymap database
There are a number of places in our code where we need to know the keycode of a
particular key. Our keymap database isn't set up for easy queries that'd let us
find these codes, and getting there would be complicated, so instead of doing
that, I extracted the codes and ranges we used into `db.constants`.

This is a curated list of keycodes Chrysalis needs, and isn't a general purpose
list.

All places that were using keycodes directly, either as keycodes, or as part of
checking whether a key falls within a range, have been updated to use the new
constants table instead.

It still has a bit of duplication, because the constants and the database must
be in sync, but as our keycodes are static, this little duplication - a well
contained one - is acceptable.

Fixes #877.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>