api/focus/macros: Add an `id` property to each macro step
For drag & drop, we need macro step ids that aren't reliant on their _current_
position in the macro array. To this end, when parsing macros, as the final
step, we add an id to each step. We don't use it for anything during
serialization, it's just there to provide a stable id.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>