FirmwareUpdate: Use a regular button with an icon, instead of SaveChangesButton
SaveChangesButton was suffering from various UX issues, such as not supporting
dark mode in disabled state, and reacting to hover events separately for the
circular and the regular part of it (in non-floating mode). The only use of it
in non-floating mode was FirmwareUpdate, which this patch changes to use a
regular Button, but still conveying similar information.

We originally used `SaveChangesButton` there because we had no other feedback
about the progress of the flashing process. We have a Stepper now that
adequately fulfills that role, so we no longer need the feedback on the button.

With the only user of non-floating `SaveChangesButton` gone, we can simplify it
greatly by removing the text part and all the workarounds that came with it.
This restores dark mode support to the button, and also makes the tooltip work
again.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>