Overhaul how Chrysalis does logging
We've been using a home-grown, pretty ad-hoc logging setup, which suffered from
a number of issues. It was simply bad, really. Log levels were all over the
place, logging was inconsistent, sometimes we logged to the console directly,
sometimes we only popped up a toast without logging it, and so on and so forth.

This is an overhaul of the whole thing.

We pull in the Winston logging library to help us out. This lets us store logs
in files too, along with logging to the developer console. We'll use this file
during the debug bundle creation, instead of keeping the session's log
in-memory.

Another big benefit of this rework is that I went through every place we log
from, and adjusted log levels to make more sense. I also added missing logging
to a number of places, so we'll get a much more useful, much more complete set
of logs.

With this new setup, we can have different log levels for the file transport,
and another for the console one. This in turn, allows us to turn down the log
level of console logs, while still having debug logs saved to disk.

Setting the log level in preferences now saves that setting, and will not reset
to the default when Chrysalis is started up again next time.

Fixes #881.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>