SaveChangesButton: Handle errors during the save
Instead of throwing an (uncaught) exception if anything goes wrong during the
save, catch it, and call the `onError` callback supplied. Because it will always
take us away from the current screen, skip any of the state changes we'd
normally do.

Both users of SaveChangesButton, MyKeyboard of Preferences and the Editor, have
been updated to handle errors gracefully.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>