main/ipc_device_discovery: Do an initial scan of devices with webusb
We're relying on webusb to send us notifications about device connect/disconnect
events, but it only sends disconnect events so for devices it knows. If there
are devices connected when we start up, we need to scam them first with webusb,
so it will know about them, and will be able to send us disconnect events.

Fixes #895.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>