MacroEditor: Set data-handler-id properly on the step Chips
For `react-dnd`, we need to set the `data-handler-id` property on the Chips, but
React does not recognise `dataHandlerId` as a component property, complains at
us about it, and renders it as `datahandlerid` - not what we want.

Instead, just use `data-handler-id` on the Chip, both Mui and React will happily
pass that along, and it will render correctly as `data-handler-id`, without
errors, or warnings.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>