api/focus: Attach an id to every request, and track it in the response
We can attach an id to every request, and because we'll be using callbacks to
return the results, we can attach the same id to the response as well. This
makes it trivial to match requests with responses.

To make that even easier, both timeout and success will now log the original
request too.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>