api/flash: Do not wait the full timeout duration for device.reset
When waiting for `device.reset` to complete, do not wait the full duration of
our normal timeout. The keyboard isn't coming back with a result if the command
succeeds.

As there is currently no way in `@api/focus` to set a custom timeout on a
per-command basis, move the reboot code to focus, and temporarily set the
timeout to 500ms for this command only.

Fixes #904.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>