api/focus: Implement our own custom parser
One major problem with `DelimiterParser` was that it only emitted data when the
delimiter was found, it was buffering meanwhile. This meant that if we wanted to
reset our timeout when partial data came in, we could not. We also built half of
a custom parser in the `data` event handler of the parser.

This patch replaces our use of `DelimiterParser` with our own, custom parser
that is explicitly tailored to the Focus protocol, and our own use case.

The new parser buffers an entire Focus response, rather than buffering line by
line. While buffering, it is running a timer, which gets reset every time we
receive data, and it gets cleared when we receive the stop marker. In practice,
this means that we no longer wait for N milliseconds for the entire response to
arrive, but N milliseconds between each partial chunk.

This will allow us to ask Kaleidoscope to flush more often when sending large
amounts of data, and then we'll be able to reduce the timeout substantially.

We also listen for a "timeout" event on the parser, and if triggered, we reject
all pending request, and call close ourselves. Similarly, we listen for "close"
events on the port, and if triggered, we cancel the parser's timer, and reject
pending requests before closing up.

With the close handler, `device.reset` will return early, and thus this patch
fixes #904 as a side effect.

And because we have an incremental timeout, also fixes #906.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>