api/log: Implement a simple log suppressor
We'd like to suppress logs if we've logged the same (sans timestamp) recently,
because there are parts of Chrysalis that can be very noisy otherwise. To this
end, implement a simple suppressor, that keeps a cache of the last few logs, and
suppresses new messages that have been seen recently. Any log that goes through
the suppressor will be added to the cache, regardless of eventual outcome, so we
always have the last N most recent ones in there.

The suppressor keeps the cache size small and tidy, defaulting to a convenient
two lines.

This, in turn, makes the `focus.find()` logs far less noisy, even on disk.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>