api/log: Implement our own browser console transport
Winston's built-in Console transport is simple, but has a major drawback: it
transports everything via `console.log`, formatted as our formatters decided to.
However, for Chrysalis, we'd like a different format on the browser console than
in files, logged at the appropriate levels, and the log info as a javascript
object, rather than stringified JSON.

The BrowserConsole transport implemented herein does just that.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>