MacroEditor: Add a step id  to newly added steps
React wants a unique key for things it handles, including macro step components.
For existing steps, we were using `step.id` as a key, but newly added steps did
not have such a property, and we ended up with an empty key, which React
complained about.

To fix this, add an id to newly created steps, and use a key for our macro steps
that can never be empty.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>