Toast: Add support for displaying an animated progress bar
For automatic upgrades, we want to show a toast when there's a new version
available. While downloading said version, we also want to show a progress bar.

To make that possible, this patch introduces `toast.progress()`, which does not
change the currently displayed toast, but updates the progress. The toast itself
will display a linear progress bar if `progress` is non-null.

Setting a new toast clears the progress, and will thus hide the bar.

When a progress bar is displayed, the close button is hidden, because we can't
cancel a download yet.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>