Implement an opt-in auto-upgrade mechanism
This replaces the old, fragile, latest version notification on the menubar with
an auto-upgrade mechanism.

While it is the main process that does the update, it is the renderer that
initiates it. We do this so that we have better control over the process, and
can display notifications on the renderer, without race conditions.

The renderer orchestrates the process via the `useAutoUpdate` hook, called by
the `App`. The hook itself will first tell the main process to check for
updates, which in turn will result in an `update-available` event being fired,
if there's an update. That event will get sent forward to the renderer as
`auto-update.update-available`, and the hook will display a toast. We handle the
rest of the upgrade process similarly, the main process pushing new
notifications to the renderer.

Because we don't want to force updates on everybody, the feature is opt-in, and
there's a new section in the User Interface tab of Preferences to enable the
feature.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>