Complete redesign of the Firmware Update screen
There were multiple complaints against the firmware update screen, and this
redesign addresses most of them. The goal was to simplify the screen, reduce the
amount of unnecessary information displayed at a time, while also showing useful
information we did not show before.

With this redesign, the update instructions are no longer shown directly on the
firmware update screen. Instead, when one selects the "Upload" or "Factory
reset" actions, the popup dialog will show the instructions instead. Previously,
clicking on "Upload" started the process immediately without confirmation - now
we ask for confirmation in this case too. Both in order to be able to display
the instructions, and to avoid accidentally triggering the process.

Then, the dropdown that was used to select the firmware to use has been replaced
by a group of radio buttons. We only ever have three choices: the current
firmware (for which a version is shown now, if available), the firmware
Chrysalis ships with, or a custom one. Having all options visible at the same
time presents useful information, and is easier to navigate than a dropdown. It
also presented a great opportunity to display the running firmware version.

Another bit of information that wasn't strictly necessary to see all the time is
the link to the firmware sources. Since we have a "View firmware changes" button
now, the link to the firmware sources was moved to the changelog dialog.

A much smaller - but still important - change is that doing a factory reset as
part of the update is now its own button, rather than a checkbox. This way it
can simply be clicked, rather than requiring ticking a checkbox and then
clicking "Update". It also makes the "Update" button's action straightforward,
it always does the same thing: updates the current firmware. It no longer
depends on external state to significantly alter its behaviour. It's only
important input is the firmware to flash.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>