FirmwareUpdate: Move the factoryreset logic to api/flash
The option to erase EEPROM during a firmware update - performing a factory reset
- has been broken for a while, and at least on the Model 100, it failed
immediately after erasing EEPROM, rendering it rather useless. To make matters
easier, lets delegate the factory reset to `api/flash`, rather than trying to
drive it from the renderer.

As an additional side-effect, the firmware update steps are now supplied by a
function, so that it can alter the steps depending on whether we're doing a
factory reset or not. This architecture will likely be useful later too, for
other purposes.

Nevertheless, the fix here revolves around handling the potential (but not
guaranteed) keyboard reboot during the erase process. Without handing it, the
firmware flashing would fail right after the wipe step, if the keyboard rebooted
as part of the erase.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>