Reapply #962, with non-linux platforms fixed
We need `/run/udev` to be able to scan devices, so if it is not accessible, or
doesn't exist, we will now display a useful error message. As this is Linux
specific, make sure that the `udev.isAvailable` IPC endpoint works on the other
platforms too, in case it is called by accident, and also don't invoke the IPC
unless we're on Linux.

This fixes #967, and reintroduces the fix for #700.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>