Move the initial theme setup into a useEffect
Because setting the theme modifies state, if we do that from the renderer
function, we need to do it from a `useEffect`. While there, move
`migrateDarkModeToTheme()` from App's renderer to index.js. We only need to do
the migration once, not on every render, so it didn't make sense to have it in
App anyway.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>