CI: Snapshot workflow improvements
Do not run the snapshot workflow for tagged releases, and when deleting the
snapshot tag during recreation, push the deletion up to the origin too.

Pushing the deletion up to the origin lets us fully recreate the tag, so it will
not point to some random old commit, but to the latest one.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>