Add static/build-info.yml
The `build-info.yml` file contains the exact version of the firmware Chrysalis
ships with, and we'll be using that version instead of Chrysalis' own version in
the very near future. As such, we need the file in `static/`.

Recent builds of Chrysalis-Firmware-Bundle bundle it up in the tarball, but for
the 0.10.4 release, it wasn't there yet, so we add it now.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>