Implement a firmware file auto update mechanism
Modeled after the application update mechanism, the firmware file auto-update
works in a similar way: if the feature is enabled (it is disabled by default),
it will check what's the latest version of the firmware on GitHub, and if it's
more recent than the version we ship with, and more recent than whatever we have
previously downloaded (if anything), then it will download the new bundle and
extract it.

The Firmware Update screen will ask the main process for the version, changelog,
and base directory to use for the default firmware. If firmware auto-update is
enabled, the main process will check first try the downloaded files, and only
fall back to the shipped version if we have nothing downloaded yet.

If firmware auto-update is disabled, then we won't even check if there's a new
version, because we'll almost always have an update, and we do not want to spam
notifications on every single startup.

Fixes #952.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>