api/focus/keymap: Update modifier-augmented key labels too when switching layouts
When switching layouts, wrap the new code table with `withModifiers()`, to add
modifier augmented keycodes to it, too. This, in turn, will make the code
following this update these modifier-augmented keycodes according to the new
table.

In practice, this means that if the host layout is set to, say, US Dvorak, then
keys like `Ctrl+Y` will render as `Ctrl+Y`, rather than `Ctrl+T` (`T` is the
same keycode on Qwerty as `Y` is on Dvorak).

Fixes #993.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>